Finish start of game
