+++ src/app/modules/game/create-game/create-game.component.html
@@ -1,26 +1,26 @@
 <form id="create-game-card" [formGroup]="gameForm" (ngSubmit)="create()">
     <div class="mb-3 input-div">
         <label for="game-id" class="form-label">Ingrese ID de juego:</label>
-        <input [attr.readonly]="raceCreated ? true : null" type="text" class="form-control" id="game-id" aria-describedby="gameId"
-            name="gameId" formControlName="gameId" required>
+        <input [attr.readonly]="readyToRace ? true : null" type="text" class="form-control" id="game-id" aria-describedby="gameId"
+            name="gameId" formControlName="gameId" required placeholder="ID Juego">
     </div>
     <div class="mb-3 input-div">
         <label for="game-id" class="form-label">Kilometros a recorrer:</label>
-        <input [attr.readonly]="raceCreated ? true : null" type="number" class="form-control" id="game-id" aria-describedby="kilometers"
+        <input [attr.readonly]="readyToRace ? true : null" type="number" class="form-control" id="game-id" aria-describedby="kilometers"
             name="kilometers" formControlName="kilometers" required>
     </div>
     <label class="form-label">Jugador 1</label>
     <div class="row mb-3 g-2">
         <div class="col">
             <div class="form-floating input-div">
-                <input [attr.readonly]="raceCreated ? true : null" type="text" class="form-control" id="player1-name" aria-describedby="player1 name"
+                <input [attr.readonly]="readyToRace ? true : null" type="text" class="form-control" id="player1-name" aria-describedby="player1 name"
                     name="playerOneName" formControlName="playerOneName" required>
                 <label for="player1-name">Nombre</label>
             </div>
         </div>
         <div class="col">
             <div class="form-floating input-div">
-                <input [attr.readonly]="raceCreated ? true : null" type="text" class="form-control" id="player1-identity" aria-describedby="player1 identification"
+                <input [attr.readonly]="readyToRace ? true : null" type="text" class="form-control" id="player1-identity" aria-describedby="player1 identification"
                     name="playerOneIdentity" formControlName="playerOneIdentity" required>
                 <label for="player1-identity">Identificación</label>
             </div>
@@ -30,14 +30,14 @@
     <div class="row mb-3 g-2">
         <div class="col">
             <div class="form-floating input-div">
-                <input [attr.readonly]="raceCreated ? true : null" type="text" class="form-control" id="player2-name" aria-describedby="player2 name"
+                <input [attr.readonly]="readyToRace ? true : null" type="text" class="form-control" id="player2-name" aria-describedby="player2 name"
                     name="playerTwoName" formControlName="playerTwoName" required>
                 <label for="player1-name">Nombre</label>
             </div>
         </div>
         <div class="col">
             <div class="form-floating input-div">
-                <input [attr.readonly]="raceCreated ? true : null" type="text" class="form-control" id="player2-identity" aria-describedby="player2 identification"
+                <input [attr.readonly]="readyToRace ? true : null" type="text" class="form-control" id="player2-identity" aria-describedby="player2 identification"
                     name="playerTwoIdentity" formControlName="playerTwoIdentity" required>
                 <label for="player1-identity">Identificación</label>
             </div>
@@ -47,20 +47,18 @@
     <div class="row mb-3 g-2">
         <div class="col">
             <div class="form-floating input-div">
-                <input [attr.readonly]="raceCreated ? true : null" type="text" class="form-control" id="player3-name" aria-describedby="player3 name"
+                <input [attr.readonly]="readyToRace ? true : null" type="text" class="form-control" id="player3-name" aria-describedby="player3 name"
                     name="playerThreeName" formControlName="playerThreeName" required>
                 <label for="player1-name">Nombre</label>
             </div>
         </div>
         <div class="col">
             <div class="form-floating input-div">
-                <input [attr.readonly]="raceCreated ? true : null" type="text" class="form-control" id="player3-identity" aria-describedby="player3 identification"
+                <input [attr.readonly]="readyToRace ? true : null" type="text" class="form-control" id="player3-identity" aria-describedby="player3 identification"
                     name="playerThreeIdentity" formControlName="playerThreeIdentity" required>
                 <label for="player1-identity">Identificación</label>
             </div>
         </div>
     </div>
-    <button *ngIf="raceCreated" class="btn btn-primary" (click)="start()">Iniciar carrera</button>
-    <button type="submit" *ngIf="!raceCreated" class="btn btn-primary">Crear Carrera</button><br>
-    <button style="margin-top: 10px;" *ngIf="!raceCreated" class="btn btn-primary" (click)="carrera()">race</button>
+    <button type="submit" *ngIf="!readyToRace" class="btn btn-primary">Crear Carrera</button>
 </form>
