+++ src/app/modules/game/create-game/create-game.component.scss
@@ -43,8 +43,12 @@ form {
     background: none;
     padding: 0.5rem 0.7rem;
     font-size: 1.2rem;
-    color: #555;
+    color: #FFFFFF;
     font-family: 'poppins', sans-serif;
+}
+
+.input-div>input::placeholder {
+    color: #d8d0d0;
 }
 
 .btn {
